refactor(mobile): split SessionCheckpoints.tsx 552->147 LOC (Cluster A2)
WHY: SessionCheckpoints packed full CRUD (fetch/save/delete/restore), a name
validator duplicated against the DB constraint, a row sub-component, and a save
modal into 552 lines - over the 400-LOC ceiling, with the validator + mapper
untestable while embedded.

Extracted into a co-located session-checkpoints/ directory:
- checkpoint-format.ts - pure validateCheckpointName + rowToCheckpoint +
                        formatCheckpointTime; now directly testable
- useSessionCheckpoints.ts - all CRUD + form state + Supabase calls
- CheckpointRow.tsx    - single-checkpoint presentational row
- SaveCheckpointModal.tsx - bottom-sheet save form (presentational)
- types.ts             - SessionCheckpointsProps
- __tests__/checkpoint-format.test.ts - 6 tests pinning the validation rules
                        (kept in lock-step with the DB length/charset/unique
                        constraints) + the snake->camel mapper

The orchestrator (147 LOC) keeps the named + default export, so the barrel and
the direct import in app/session/[id].tsx are unchanged. This completes the
mobile side of Cluster A2 (5 components split; web's 2 remain).

Verification: mobile typecheck + lint clean; full suite 1753 pass (+6 new);
a11y regression guard green (accessibility props preserved on all touchables).
